+++ sensorManagerExample/src/main/java/com/ubhave/sensormanager/tester/pull/SensorConfigUpdater.java
@@ -262,7 +262,7 @@ public class SensorConfigUpdater extends ExampleSensorConfigUpdater
 	}
 
 	@Override
-	public void setTimeThreshold(int millis)
+	public void setTimeThreshold(long millis)
 	{
 		super.setTimeThreshold(millis);
